+++ fort-to-nite/Models/Weapons.swift
@@ -15,6 +15,7 @@ class Weapons: Decodable {
     var color: Int
     var img: String
     var group: Int
+    var is_removed: Int
     
     init() {
         id = 0
@@ -22,14 +23,16 @@ class Weapons: Decodable {
         color = 0
         img = ""
         group = 0
+        is_removed = 0
     }
     
-    init(id: Int, name: String, color: Int, img: String, group: Int) {
+    init(id: Int, name: String, color: Int, img: String, group: Int, is_removed: Int) {
         self.id = id
         self.name = name
         self.color = color
         self.img = img
         self.group = group
+        self.is_removed = is_removed
     }
     
 }